// src/shared/components/SvgSwitch/SvgSwitch.types.ts
export interface ISvgSwitchProps {
    svg: SVG;
    testID?: string;
};

type SVG = "arrow"
           | "chemist"
           | "click"
           | "dollar"
           | "electric"
           | "globe"
           | "hand"
           | "laptop"
           | "pencil"
           | "phone"
           | "pills"
           | "plane"
           | "pound"
           | "spanner"
           | "tshirt"
           | "passport"
           | "license";